Add ability to build trigger to check job status
This commits adds the ability to the build trigger to check the job
status which can used to track the status of the individual jobs in
the pipeline to report failures which we can use to stop the trigger
from proceeding with the cluster rebuild. This way we can investigate
what went wrong. If the job.status file is not found or set to true,
the cluster rebuild just depends on signal obtained from comparing
current vs previous builds.